+++ src/util.h
@@ -47,7 +47,7 @@ void scope_extract_back( char* scope, char* back, char* rest );
 bool scope_local( char* scope );
 
 //! Performs safe malloc call.
-void* malloc_safe( int size );
+void* malloc_safe(size_t size );
 
 //! Performs safe deallocation of heap memory.
 void free_safe( void* ptr );
@@ -56,6 +56,10 @@ void free_safe( void* ptr );
 void gen_space( char* spaces, int num_spaces );
 
 /* $Log$
+/* Revision 1.5  2002/07/18 22:02:35  phase1geo
+/* In the middle of making improvements/fixes to the expression/signal
+/* binding phase.
+/*
 /* Revision 1.4  2002/07/08 12:35:31  phase1geo
 /* Added initial support for library searching.  Code seems to be broken at the
 /* moment.
